physics basic implementation and drawing

+++ GameEngine/KekEngine/includes/livingEntity.h
@@ -28,8 +28,8 @@ public:
     void    changeDirection(const LivingEntity::Direction &direction);
     void    move(const float &x, const float &y);
     void    setPosition(const double &x, const double &y);
-    virtual void    setY(const double &y);
-    virtual void    setX(const double &x);
+    virtual void    setY(const float &y);
+    virtual void    setX(const float &x);
     void    update(const sf::Time &time);
     const float &getSpeed() const;
     AnimatedSprite* getAnimation() const;
